pyproject.toml
[project]
name = "chronofact-ai"
version = "0.1.0"
description = "AI-powered fact-based news service that ethically ingests data from X, constructs accurate and verifiable event timelines using Qdrant and BAML"
authors = [
    { name = "Ayush4513", email = "sharmayush045@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "baml-py>=0.61.0",
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "typer>=0.12.0",
    "selenium>=4.20.0",
    "webdriver-manager>=4.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
    # Multimodal support
    "Pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
sources = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
